chore: moved serialize interceptor into the core group
