+++ src/orders/orders.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Delete, Get, Post, UseGuards } from '@nestjs/common';
 import { AuthGuard } from 'src/core/guards/auth.guard';
-import { Serialize } from 'src/interceptors/serialize.inteceptor';
+import { Serialize } from 'src/core/interceptors/serialize.inteceptor';
 import { CurrentUser } from 'src/core/decorators/current-user.decorators';
 import { CreateUserDTO } from 'src/users/dtos/dto';
 import { CreateOrderDTO } from './dto/create-order.dto';
